+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- AreaTriggerUnityEvent component
 - AreaTrigger component
 - Attributes package dependency
 - CollisionableDetector component